+++ app/src/main/java/xyz/thingapps/rssliveslider/adapters/ItemListAdapter.kt
@@ -3,7 +3,7 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import xyz.thingapps.rssliveslider.R
-import xyz.thingapps.rssliveslider.api.dao.Item
+import xyz.thingapps.rssliveslider.api.Item
 import xyz.thingapps.rssliveslider.viewholders.ItemViewHolder
 
 class ItemListAdapter(var currentChannel: Int, private val tag: Int) :
