+++ app/src/main/java/xyz/thingapps/rssliveslider/api/RssApi.kt
@@ -3,7 +3,6 @@ package xyz.thingapps.rssliveslider.api
 import io.reactivex.Observable
 import retrofit2.http.GET
 import retrofit2.http.Url
-import xyz.thingapps.rssliveslider.api.dao.Cast
 
 interface RssApi {
     @GET
